requirements.txt
# Async HTTP and WebSocket
aiohttp>=3.9.0
websockets>=12.0

# Environment management
python-dotenv>=1.0.0

# HTTP client (sync, for simpler REST calls)
httpx>=0.27.0

# Data handling
orjson>=3.9.0

# Type hints runtime (optional but useful)
typing-extensions>=4.9.0